Desconto

+++ trab1/src/main/java/br/unb/model/Produto.java
@@ -4,10 +4,10 @@ public class Produto {
 
     public final String descricao;
     public final String codigo;
-    public final float valorDeVenda;
+    public final double valorDeVenda;
     public final String unidade;
 
-    public Produto(String descricao, float valorDeVenda, String unidade, String codigo) {
+    public Produto(String descricao, double valorDeVenda, String unidade, String codigo) {
         this.descricao = descricao;
         this.codigo = codigo;
         this.valorDeVenda = valorDeVenda;
